+++ pipeline_inspection/pipeline_inspection/follow_pipeline.py
@@ -4,50 +4,9 @@ import rclpy
 from rclpy.node import Node
 import threading
 
-from mavros_wrapper.ardusub_wrapper import *
-
 from rcl_interfaces.msg import ParameterType
 from pipeline_inspection_msgs.srv import GetPath
-from geometry_msgs.msg import Pose
-from std_msgs.msg import Bool
-
-
-class BlueROVGazeboNode(BlueROVArduSubWrapper):
-    def __init__(self, node_name='bluerov_gz_mavros'):
-        super().__init__(node_name)
-        self.gazebo_pos_sub = self.create_subscription(
-            Pose, 'model/bluerov2/pose', self.gazebo_pos_cb, 10)
-        self.first_gazebo_pos_msg = True
-
-        # TODO: make this a ros param
-        self.altitude = 2
-
-        self.start_follow = False
-
-        self.start_follow_sub = self.create_subscription(
-            Bool, 'start_follow_pipe', self.follow_pipe_cb, 10)
-
-    def gazebo_pos_cb(self, msg):
-        self.gazebo_pos = msg
-        if self.first_gazebo_pos_msg and self.local_pos_received:
-            self.gz_local_delta_pos = [
-                self.local_pos.pose.position.x - msg.position.x,
-                self.local_pos.pose.position.y - msg.position.y,
-                self.local_pos.pose.position.z - msg.position.z,
-            ]
-            self.first_gazebo_pos_msg = False
-
-    def convert_gz_to_local_pos(self, gz_pose):
-        local_pose = Pose()
-        local_pose.position.x = gz_pose.position.x + self.gz_local_delta_pos[0]
-        local_pose.position.y = gz_pose.position.y + self.gz_local_delta_pos[1]
-        local_pose.position.z = \
-            gz_pose.position.z + self.gz_local_delta_pos[2] + self.altitude
-
-        return local_pose
-
-    def follow_pipe_cb(self, msg):
-        self.start_follow = msg.data
+from pipeline_inspection.bluerov_gazebo import BlueROVGazebo
 
 
 def mission(ardusub):
@@ -85,7 +44,7 @@ def main():
     # Initialize ros node
     rclpy.init(args=sys.argv)
 
-    ardusub = BlueROVGazeboNode("follow_pipeline_node")
+    ardusub = BlueROVGazebo("follow_pipeline_node")
 
     thread = threading.Thread(target=rclpy.spin, args=(ardusub, ), daemon=True)
     thread.start()
